fix: changed simple types

+++ WeatherWinFormsApp/WeatherWinFormsApp/API/CommonTypes.cs
@@ -28,7 +28,7 @@ namespace WeatherWinFormsApp.API
     {
         public float temp;
         public float feels_like;
-        public int temp_min;
+        public float temp_min;
         public float temp_max;
         public int pressure;
         public int humidity;
